Made EditorModal more reusable for multi-stage modals
